[TASK] Do not use full TYPO3 context in early permission checks
`PermissionsCheck` verifies whether the configured database
has the required permissions. For this purpose, it is not
necessary to use TYPO3's Connection and QueryBuilder.

Using QueryBuilder too early can lead to issues, since it
applies default restrictions that rely on TCA and the
TcaSchemaFactory, which may not be available during early
initialization.

This change replaces QueryBuilder with a plain Doctrine
connection for simple permission checks, avoiding unnecessary
dependencies on TYPO3 internals at this stage.

Resolves: #106999
Releases: main, 13.4
Change-Id: Icd9e5b8c3d540972ca80e46ae494f46749884961
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89937
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>