[BUGFIX] Make Install Tool AJAX error handling resilient to JS errors
If (for some reason) there's a JavaScript error while doing
an AJAX action in an Install Tool module modal, it will just
end up freezing because `error.response` is not set, as
there potentially never was a response.

This change adds null-checks to the response property in the
AJAX error handling and also shows the JS error message in the
"Ajax error" collapsible to inform the user about the issue.

Resolves: #82287
Releases: main, 13.4
Change-Id: Ia1c83f85c5329c181dd473c50d2001a2dd96e8c0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90181
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>