[!!!][TASK] Remove random subpage option of doktype=shortcut
The functionality of "Random subpage of current / selected page"
when choosing a page of type "shortcut" was removed
from TYPO3 Core.

It had little use and wasn't really "random" as the random
value was cached for pages that had links to this page,
as the links are resolved since TYPO3 v11 on page generation
and not during resolution.

Resolves: #107654
Releases: main
Change-Id: I0ea6c52839a0e02b7e4ca7327c54b57de50199e7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90993
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>