[TASK] Convert registerExtractionService deprecation to breaking
The deprecation of ExtractorRegistry::registerExtractionService()
has been converted to a breaking change to allow extensions to be
compatible with both TYPO3 v13 and v14 without version checks.

The method is now a no-op (trigger_error removed) and the
documentation has been updated from Deprecation to Breaking,
following the same pattern as other hook-to-event and
auto-configuration migrations.

This allows extensions to call the method in both versions:
- v13: Registration works as expected
- v14+: Method does nothing, registration via interface

Resolves: #107897
Related: #107786
Releases: main
Change-Id: Icf1bd50539884c34c006f0adc0f28f830c8d6451
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91319
Tested-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>