[BUGFIX] Do not delete history in install tool + scheduler
One of the strenghts of TYPO3 being labeled as
Enterprise CMS, the history and audit trail is
quite powerful.

However, in the past, sys_history and sys_log
was intermingled (for e.g. newly created records etc),
so this was extracted in v9. in v13 we've added
proper workspaces support for sys_history, in v14
scheduler tasks have been tracked.

We should - by any means - discourage all admins
to not "simply" truncate sys_history.

Further solutions can be evaluated on how to thin
out history state, but for now, we discourage
this via UI.

Resolves: #107903
Releases: main, 13.4
Change-Id: I7898ec959d16e98abe77c0e4bb75c957e60f8421
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91326
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Josef Glatz <typo3@josefglatz.at>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oli Bartsch <bo@cedev.de>