[BUGFIX] Have simple cropper VH in ext:install
We made ext:fluid CropViewHelper DI aware in #108161
but missed it is used in ext:install where container
is not fully loaded, leading to fatals in 'Configure
Installation-Wide Options'.

We could hard code DI config for this case, but ext:install
uses a fraction of the VH functionality only, it is easier
to just use a more slim own version of that view helper
instead.

Note we may end up adding basic string cropping to fluid
standalone directly, so this view helper may also vanish
later again.

Resolves: #108202
Related: #108161
Releases: main
Change-Id: Ieed8f3e187dd210dd92e8eda906d173299dde4f6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91764
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Garvin Hicking <garvin@hick.ing>