[BUGFIX] Fix Extension Scanner crash on dynamic method calls
Several matchers in the Extension Scanner crash when encountering
dynamic method calls like `$object->$methodName()`. This happens
because the code assumes `$node->name` is always an identifier with
a 'name' property, but for dynamic calls it can be a variable.

This is fixed by adding an explicit type check using
an `instanceof` check before accessing `$node->name->name`.
This follows the pattern already used correctly
in `MethodArgumentUnusedMatcher`.

Added tests fixate this behavior.

Affected matchers:
- MethodArgumentDroppedMatcher
- MethodArgumentDroppedStaticMatcher
- MethodArgumentRequiredMatcher
- MethodArgumentRequiredStaticMatcher
- MethodArgumentUnusedMatcher
- MethodCallMatcher
- MethodCallStaticMatcher
- InterfaceMethodChangedMatcher
- PropertyExistsStaticMatcher

Resolves: #108413
Releases: main, 14.3
Change-Id: I365cfed93e1cbd5016c4ba32c96c7051a7dad7a9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/94497
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>